Add 'FeedImageViewPresenterInput' protocol for abstraction between controller and presenter
