+++ 3.EssentialFeedProject/EssentialFeedProject/EssentialFeed/EssentialFeediOS/Feed UI/Controller/FeedImageCellController.swift
@@ -8,11 +8,13 @@
 import Foundation
 import UIKit
 
-final class FeedImageCellController: FeedImageView {
-    private let presenter: FeedImageViewPresenter<FeedImageCellController, UIImage>
+final class FeedImageCellController {
+    typealias Image = UIImage
+    
+    private let presenter: FeedImageViewPresenterInput
     private var imageCell: FeedImageCell?
     
-    init(presenter: FeedImageViewPresenter<FeedImageCellController, UIImage>){
+    init(presenter: FeedImageViewPresenterInput) {
         self.presenter = presenter
     }
     
@@ -23,25 +25,27 @@ final class FeedImageCellController: FeedImageView {
         return view
     }
     
-    func display(_ viewModel: FeedImageViewModel<UIImage>) {
-        guard let view = imageCell else { return }
-        load(view, with: viewModel)
-    }
-    
-    private func load(_ view: FeedImageCell, with viewModel: FeedImageViewModel<UIImage>) {
-        view.imageDescription = viewModel.description
-        view.location = viewModel.location
-        view.feedImageView.image = viewModel.image
-        view.retryImageLoad.isHidden = !viewModel.shouldRetry
-        view.retryImageAction = presenter.loadImage
-        view.locationContainer.isHidden = !viewModel.hasLocation
-    }
-    
     func preload() {
         presenter.loadImageData()
     }
     
     func cancelTask() {
         presenter.cancelImageData()
+    }
+}
+
+extension FeedImageCellController: FeedImageView {
+    func display(_ viewModel: FeedImageViewModel<UIImage>) {
+        guard let view = imageCell else { return }
+        load(view, with: viewModel)
+    }
+    
+    private func load(_ view: FeedImageCell, with viewModel: FeedImageViewModel<Image>) {
+        view.imageDescription = viewModel.description
+        view.location = viewModel.location
+        view.feedImageView.image = viewModel.image
+        view.retryImageLoad.isHidden = !viewModel.shouldRetry
+        view.retryImageAction = presenter.loadImage
+        view.locationContainer.isHidden = !viewModel.hasLocation
     }
 }
